+++ CHANGELOG.md
@@ -2,7 +2,7 @@
     - Switch back to original fork of Tensor
     - Added `maxBins` hyper-parameter to CART-based learners
     - Added stream Deduplicator extractor
-    - Added the Swish activation function
+    - Added the SiLU activation function
 
 - 1.2.2
     - Allow empty dataset objects in `stack()`